# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Shark server."
  version: "1.0.0"
  title: "Shark API"
tags:
  - name: "shark"
    description: "King of the beach"
schemes:
  - "http"
paths:
  /sharks:
    get:
      tags:
        - "shark"
      summary: "Find all shark"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Shark"
    post:
      tags:
        - "shark"
      summary: "Add a new shark"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/SharkCommand"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Shark"
        "400":
          description: "bad request"
    put:
      tags:
        - "shark"
      summary: "Update an existing shark"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/SharkCommand"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Shark"
        "400":
          description: "bad request"
        "404":
          description: "not found"
  /sharks/{sharkId}:
    get:
      tags:
        - "shark"
      summary: "Find shark by ID"
      produces:
        - "application/json"
      parameters:
        - name: "sharkId"
          in: "path"
          description: "ID of shark to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Shark"
        "400":
          description: "bad request"
        "404":
          description: "not found"
    delete:
      tags:
        - "shark"
      summary: "Deletes a shark"
      produces:
        - "application/json"
      parameters:
        - name: "sharkId"
          in: "path"
          description: "shark id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "400":
          description: "bad request"
        "404":
          description: "not found"
definitions:
  Shark:
    type: "object"
    required:
      - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
      name:
        type: "string"
        example: "jaws"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
  SharkCommand:
    type: "object"
    required:
      - "name"
    properties:
      name:
        type: "string"
        example: "jaws"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"